refactor(runtime): split session lifecycle from runtime mod
